# backend/openapi.yaml
openapi: 3.0.3
info:
  title: VA Matching Platform API
  description: API for matching Virtual Assistants with Companies
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.vaplatform.com
    description: Production server
paths:
...
  /va/profile:
    get:
      summary: Get VA profile
      tags:
        - VA Profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: VA profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VAProfile'
        '404':
          description: Profile not found
    post:
      summary: Create VA profile
      tags:
        - VA Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VAProfileCreate'
      responses:
        '201':
          description: VA profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VAProfile'
        '400':
          description: Validation error
    put:
      summary: Update VA profile
      tags:
        - VA Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VAProfileUpdate'
      responses:
        '200':
          description: VA profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VAProfile'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [company, va]
        createdAt:
          type: string
          format: date-time
    VAProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        country:
          type: string
        hourlyRate:
          type: integer
        skills:
          type: array
          items:
            type: string
        availability:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
    VAProfileCreate:
      type: object
      required:
        - name
        - country
        - hourlyRate
        - skills
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        country:
          type: string
          minLength: 2
          maxLength: 50
        hourlyRate:
          type: integer
          minimum: 1
          maximum: 200
        skills:
          type: array
          items:
            type: string
          minItems: 1
        availability:
          type: boolean
          default: true
    VAProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        country:
          type: string
          minLength: 2
          maxLength: 50
        hourlyRate:
          type: integer
          minimum: 1
          maximum: 200
        skills:
          type: array
          items:
            type: string
        availability:
          type: boolean
...